+++ Code/RoboX/map_core.h
@@ -20,6 +20,7 @@
 #include "Arduino.h"
 #include "config.h"
 #include <inttypes.h>
+#include <stdlib.h>
 
 
 /// DEFINES ///
@@ -29,8 +30,8 @@
 
 /// STRUCTS ///
 typedef struct {
-  uint8_t x;
-  uint8_t y;
+  uint16_t x;
+  uint16_t y;
 } Position_t;
